+++ CHANGELOG.md
@@ -14,3 +14,9 @@
 ## 1.0.5
 
 * Redundant code removed and the code has been optimized.
+
+## 1.0.6
+
+* Added `FastCachedVideo` widget to cache and play videos.
+* Added `FastCachedLottie` widget to cache and play lottie animations.
+* Updated dependencies.